+++ src/types.ts
@@ -1,0 +1,15 @@
+interface ArticleSource {
+    id: string;
+    name: string;
+}
+
+export interface Article {
+    source: ArticleSource;
+    author: string | null;
+    title: string;
+    description: string;
+    url: string;
+    urlToImage: string | null;
+    publishedAt: string;
+    content: string;
+}
